fixed bug in asm_list
when creating instructions for an error condition, the context is
setting a variable. the variable is meant to be the result of an
if expression, however the error condition has no intention of
returning anything from the if expression as it is calling system
exit. this would be fine except for the fact that the live list
calculation will see the read of the result of that branch of the
if expression outside of the expression, but will never see that
result written to, thus the live list calculation will accumulate
useless variables and things get bad :(