added tracer class to track the generations of nodes
had to make some of the ASTTxformers somewhat more consistent
on their return values to make the tracer straighforward.
basically the tracer sees the node that was dispatched, and
then sees the nodes generated from the dispatched node. it
sets the dispatched node as the parent of all the generated
nodes. using this we should be able to trace any AST node back
to the source code line that generated it.